Updated kernel tree - bq27x00 debug stuff

+++ arch/um/os-Linux/time.c
@@ -8,9 +8,11 @@
 #include <signal.h>
 #include <time.h>
 #include <sys/time.h>
+#include "kern_constants.h"
 #include "kern_util.h"
 #include "os.h"
-#include "internal.h"
+#include "process.h"
+#include "user.h"
 
 int set_interval(void)
 {
@@ -58,7 +60,7 @@ static inline long long timeval_to_ns(const struct timeval *tv)
 long long disable_timer(void)
 {
 	struct itimerval time = ((struct itimerval) { { 0, 0 }, { 0, 0 } });
-	long long remain, max = UM_NSEC_PER_SEC / UM_HZ;
+	int remain, max = UM_NSEC_PER_SEC / UM_HZ;
 
 	if (setitimer(ITIMER_VIRTUAL, &time, &time) < 0)
 		printk(UM_KERN_ERR "disable_timer - setitimer failed, "
